Allow custom controllers for steps (#419)
* Allow custom controllers for steps

* move jurisdiction controller into app
