+++ src/main/app/jurisdiction/JurisdictionPostController.test.ts
@@ -1,12 +1,12 @@
-import { mockRequest } from '../../../../test/unit/utils/mockRequest';
-import { mockResponse } from '../../../../test/unit/utils/mockResponse';
-import { JurisdictionConnections, PATCH_CASE, YesOrNo } from '../../../app/case/definition';
-import { Form } from '../../../app/form/Form';
+import { mockRequest } from '../../../test/unit/utils/mockRequest';
+import { mockResponse } from '../../../test/unit/utils/mockResponse';
+import { JurisdictionConnections, PATCH_CASE, YesOrNo } from '../case/definition';
+import { Form } from '../form/Form';
 
+import { JurisdictionPostController } from './JurisdictionPostController';
 import { addConnection } from './connections';
-import { JurisdictionPostController } from './post';
 
-jest.mock('../../../steps/jurisdiction/interstitial/connections');
+jest.mock('./connections');
 const addConnectionMock = addConnection as jest.Mock<JurisdictionConnections[]>;
 
 describe('JurisdictionPostController', () => {
