+++ src/main/steps/living-england-wales-six-months/post.ts
@@ -1,0 +1,3 @@
+import { JurisdictionPostController } from '../../app/jurisdiction/JurisdictionPostController';
+
+export default JurisdictionPostController;
