+++ src/main/steps/where-your-lives-are-based/post.ts
@@ -1,0 +1,3 @@
+import { JurisdictionPostController } from '../../app/jurisdiction/JurisdictionPostController';
+
+export default JurisdictionPostController;
