+++ src/main/steps/your-domicile/post.ts
@@ -1,0 +1,3 @@
+import { JurisdictionPostController } from '../../app/jurisdiction/JurisdictionPostController';
+
+export default JurisdictionPostController;
